requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.36
asyncpg==0.30.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4